requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23
SQLAlchemy>=2.0
pydantic>=2.7
pydantic-settings>=2.3
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
pydantic[email]>=2.7

